Start Parser
I wanted to see what ChatGPT could get started for me, so I gave it my lexer program and asked for a recursive descent parser. It did not disappoint all of parser.cpp was written by it. Please refer to my project 2 documentation to see my narration of this project.