Got a bunch done
So I realized a number with multiple digits is not accepted. I have to fix that and there is a bunch more TODOs i added. In this commit I made the CST and fixed parsing multiple programs. I have to leave for my spring break trip soon, so hopefully I finish the rest of the project but at this point id estimate this is like a 80 (id hope).